Back/requirements.txt
Flask
redis
boto3
pymysql
python-dotenv
